add x509-cert-signer-keyid and ssh-cert-signer-keyid fields to domain meta
