+++ servers/zms/schema/updates/update-20240708.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `zms_server`.`domain` ADD `x509_cert_signer_keyid` VARCHAR(64) NOT NULL DEFAULT '';
+ALTER TABLE `zms_server`.`domain` ADD `ssh_cert_signer_keyid` VARCHAR(64) NOT NULL DEFAULT '';
