+++ servers/zms/schema/zms_server.sql
@@ -1,5 +1,5 @@
 -- MySQL Script generated by MySQL Workbench
--- Fri May 24 13:53:41 2024
+-- Mon Jul  8 11:11:29 2024
 -- Model: New Model    Version: 1.0
 -- MySQL Workbench Forward Engineering
 
@@ -53,6 +53,8 @@ CREATE TABLE IF NOT EXISTS `zms_server`.`domain` (
   `resource_owner` VARCHAR(256) NOT NULL DEFAULT '',
   `azure_tenant` VARCHAR(128) NOT NULL DEFAULT '',
   `azure_client` VARCHAR(128) NOT NULL DEFAULT '',
+  `x509_cert_signer_keyid` VARCHAR(64) NOT NULL DEFAULT '',
+  `ssh_cert_signer_keyid` VARCHAR(64) NOT NULL DEFAULT '',
   PRIMARY KEY (`domain_id`),
   UNIQUE INDEX `uq_name` (`name` ASC),
   INDEX `idx_modified` (`modified` ASC),
